requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.9.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==25.3.0
    # via -r requirements/base.txt
casbin-django-orm-adapter==1.7.0
    # via -r requirements/base.txt
certifi==2025.8.3
    # via
    #   -r requirements/base.txt
    #   requests
cffi==2.0.0
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.4.3
    # via
    #   -r requirements/base.txt
    #   requests
click==8.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==2.3.0
    # via -r requirements/test.in
coverage[toml]==7.10.6
    # via pytest-cov
cryptography==46.0.2
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   casbin-django-orm-adapter
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-django-utils
    #   edx-drf-extensions
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-waffle==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.16.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-drf-extensions
dnspython==2.8.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
drf-yasg==1.21.11
    # via
    #   -r requirements/base.txt
    #   edx-api-doc-tools
edx-api-doc-tools==2.1.0
    # via -r requirements/base.txt
edx-django-utils==8.0.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==10.6.0
    # via -r requirements/base.txt
edx-opaque-keys==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
openedx-atlas==0.7.0
    # via -r requirements/base.txt
packaging==25.0
    # via
    #   -r requirements/base.txt
    #   drf-yasg
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psutil==7.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycasbin==2.2.0
    # via
    #   -r requirements/base.txt
    #   casbin-django-orm-adapter
    #   redis-watcher
pycparser==2.23
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.10.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.15.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.6.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.4.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pytz==2025.2
    # via
    #   -r requirements/base.txt
    #   drf-yasg
pyyaml==6.0.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   drf-yasg
redis==6.4.0
    # via
    #   -r requirements/base.txt
    #   redis-watcher
redis-watcher==1.8.0
    # via -r requirements/base.txt
requests==2.32.5
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
simpleeval==1.0.3
    # via
    #   -r requirements/base.txt
    #   pycasbin
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.5.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
uritemplate==4.2.0
    # via
    #   -r requirements/base.txt
    #   drf-yasg
urllib3==2.5.0
    # via
    #   -r requirements/base.txt
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
